Add files via upload
This is an update to DiceStats that actually works and runs.  Please note that it has not been implemented into the app navigation and therefore can't be navigated to yet.  This means it may require some minor testing to make sure it's 100%, but while it's a lot of code, it's relatively simple in implementation, so I'm not expecting any serious issues.